Order and cart updates to main

+++ backend/product-service/src/main/java/com/buy01/product/service/ProductService.java
@@ -51,7 +51,8 @@ public class ProductService {
     public ProductResponseDTO createProduct(ProductCreateDTO request, AuthDetails currentUser) throws IOException {
 
         // validate that user can create products
-        if (currentUser.getCurrentUserId().isEmpty() || (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getRole().equals(Role.SELLER))) {
+        if (currentUser.getCurrentUserId().isEmpty() ||
+            (!currentUser.getRole().equals(Role.ADMIN) && !currentUser.getRole().equals(Role.SELLER))) {
             log.info("Forbidden: User ID is empty or role is not allowed - Role: {}", currentUser.getRole());
             throw new ForbiddenException("Your current role cannot create a product.");
         }
@@ -211,24 +212,39 @@ public class ProductService {
     }
 
     // Update product quantity, called when product quantity is changed in the cart
-    public void updateProductQuantity(String productId, int requestedQuantity) {
+    public void updateProductQuantity(String productId, int delta) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new NotFoundException(productId));
-        int newQuantity = product.getQuantity() + requestedQuantity;
+
+        int newQuantity = product.getQuantity() + delta;
+
         if (newQuantity < 0) {
             throw new ConflictException("Insufficient product quantity, remaining quantity" + product.getQuantity());
         }
+        updateReservedQuantity(product, -delta);
         product.setQuantity(newQuantity);
         product.setUpdateTime(new Date());
         productRepository.save(product);
 
-        // send productUpdate via Kafka
-        productEventService.publishProductUpdatedEvent(new ProductUpdateDTO(
-                product.getProductId(),
-                product.getName(),
-                product.getPrice(),
-                product.getQuantity()
-        ));
+    }
+
+    public void removeReserveQuantityForOrderPlaced(String productId, int delta) {
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> new NotFoundException(productId));
+        updateReservedQuantity(product, delta);
+        product.setUpdateTime(new Date());
+        productRepository.save(product);
+    }
+
+    public void updateReservedQuantity(Product product, int delta) {
+
+        int newReservedQuantity = product.getReservedQuantity() + delta;
+
+        if (newReservedQuantity < 0 ) {
+            throw new ConflictException("Invalid reserved quantity operation.");
+        }
+
+        product.setReservedQuantity(newReservedQuantity);
     }
 
     // Deleting product, accessible only by ADMIN or product owner
@@ -237,6 +253,9 @@ public class ProductService {
                .orElseThrow(() -> new NotFoundException(productId));
 
         authProductOwner(product, currentUser.getCurrentUserId(), currentUser.getRole());
+        if (product.getReservedQuantity() > 0) {
+            throw new ConflictException("Cannot delete product that is in the shopping cart with quantity: " + product.getReservedQuantity());
+        }
 
         productRepository.deleteById(productId);
         productEventService.publishProductDeletedEvent(productId);
@@ -254,19 +273,6 @@ public class ProductService {
     }
 
     // Helper methods
-
-    // Validate product details
-    private void validateProduct(Product product) {
-        if (product.getName() == null || product.getName().isEmpty()) {
-            throw new IllegalArgumentException("Product name is required");
-        }
-        if (product.getPrice() == null) {
-            throw new IllegalArgumentException("Price is required");
-        }
-        if (product.getPrice() <= 0) {
-            throw new IllegalArgumentException("Price must be positive");
-        }
-    }
 
     private void validateProductName(String productName) {
         if (productName == null || productName.isBlank()) {
